Bugfix/Memory: Replace std::array with standard C-style array
I was unaware that std::array allocated to the stack by default.
Replace it with something more elegant later, for now this is an
acceptable fix.

We now have to zero the general purpose registers for some reason.
Might add a toggle for this later, since real processors have
randomized register contents on boot.