Cleanup: Dropping down the verbosity a bit.
A new variable called 'branched' gets set to true if encountering a branch or jump instruction.
If 'branched' is false by the end of the loop, the program counter is incremented 4.

Moved commonly used variables out of the switch chain.

Deleted all calls to unimplemented instruction, the new pretty decodes will handle it in the future.